Return reader, writer pair instead of EscVp21Communication
Also removes the contextmanager
